feature(Notifications): initial documentation
