+++ DidIFloss/DidIFlossApp.swift
@@ -22,7 +22,7 @@ struct DidIFlossApp: App {
 
 class AppDelegate: NSObject, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-        NotificationManager.requestAuthorizationToNotificate(provisional: true)
+        Notification.Service.requestAuthorizationToNotificate(provisional: true)
         
         return true
     }
