+++ DidIFloss/Service/Persistence/Extensions/UserDefaults.swift
@@ -6,3 +6,9 @@
 //
 
 import Foundation
+
+/// An extension of UserDefaults adopting the UserDefaultable.
+///
+/// This extension makes UserDefaults conform to UserDefaultable,
+/// allowing you to use UserDefaults or any custom implementation conforming to UserDefaultsProtocol interchangeably.
+extension UserDefaults: UserDefaultable {}
